Fixing the Bellman's maximizing algorithme: detected some bugs while executing test_dijkstra_maximize, for comparison with bellman maximisation

+++ inc/GraphTypes.hpp
@@ -30,8 +30,8 @@ public:
 
   enum OptimizationType{MAXIMIZE, MINIMIZE};
 
-  static const int POSITIVE_INFINITY;
-  static const int NEGATIVE_INFINITY;
+  static const float POSITIVE_INFINITY;
+  static const float NEGATIVE_INFINITY;
 };
 
 #endif
